Create Map the Debris
This one got ugly, but I'm pretty proud of it.  My biggest source of frustration was the inaccuracy from using only the first 5 digits of pi, which caused one of the calculations to be off by an entire two minutes; Using the first 60 digits of pi instead solved this beautifully.