function complete
Fixed both duplicate values and duplicate index problems, with the help of an additional array with the values of the first array.  When a pair is found, those two numbers are changed in the second array and replaced with an "X" string.  The next instance can't find a number in that index spot any more, and so skips it.